requirements.txt
adal==1.2.7
adlfs==2024.7.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
apache-airflow==2.9.3
apache-airflow-providers-celery==3.7.2
apache-airflow-providers-common-io==1.3.2
apache-airflow-providers-common-sql==1.14.2
apache-airflow-providers-fab==1.2.1
apache-airflow-providers-ftp==3.10.0
apache-airflow-providers-http==4.12.0
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-microsoft-azure==10.2.0
apache-airflow-providers-microsoft-mssql==3.7.2
apache-airflow-providers-smtp==1.7.1
apache-airflow-providers-sqlite==3.8.1
apispec==6.6.1
argcomplete==3.4.0
asgiref==3.8.1
attrs==23.2.0
azure-batch==14.2.0
azure-common==1.1.28
azure-core==1.30.2
azure-cosmos==4.7.0
azure-datalake-store==0.0.53
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
azure-kusto-data==4.5.1
azure-mgmt-containerinstance==10.1.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-core==1.4.0
azure-mgmt-cosmosdb==9.5.1
azure-mgmt-datafactory==8.0.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-resource==23.1.1
azure-mgmt-storage==21.2.1
azure-nspkg==3.0.2
azure-servicebus==7.12.2
azure-storage-blob==12.21.0
azure-storage-file-datalake==12.16.0
azure-storage-file-share==12.17.0
azure-synapse-artifacts==0.19.0
azure-synapse-spark==0.7.0
Babel==2.15.0
billiard==4.2.0
blinker==1.8.2
cachelib==0.9.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.3
croniter==2.0.5
cryptography==41.0.7
Deprecated==1.2.14
dill==0.3.1.1
dnspython==2.6.1
docutils==0.16
email_validator==2.2.0
et-xmlfile==1.1.0
Flask==2.2.5
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flower==2.0.1
frozenlist==1.4.1
fsspec==2023.12.2
google-re2==1.1.20240702
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.1
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
humanize==4.10.0
hyperframe==6.0.1
idna==3.7
ijson==3.3.0
importlib-metadata==6.11.0
importlib_resources==6.4.0
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
lazy-object-proxy==1.10.0
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.1
mdurl==0.1.2
methodtools==0.4.7
microsoft-kiota-abstractions==1.3.3
microsoft-kiota-authentication-azure==1.0.0
microsoft-kiota-http==1.3.2
more-itertools==10.3.0
msal==1.30.0
msal-extensions==1.2.0
msgraph-core==1.1.2
msrest==0.7.1
msrestazure==0.6.4.post1
multidict==6.0.5
numpy==2.0.1
oauthlib==3.2.2
openpyxl==3.1.5
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
ordered-set==4.1.0
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pendulum==3.0.0
pluggy==1.5.0
portalocker==2.10.1
prison==0.2.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
pymssql==2.3.0
pyodbc==5.1.0
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.7.1
rich-argparse==1.5.2
rpds-py==0.19.0
setproctitle==1.3.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.52
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.0
std-uritemplate==1.0.5
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.14.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.2
urllib3==1.26.19
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.4
zipp==3.19.2
